+++ libs/cart/src/injection-tokens/public_api.ts
@@ -1,3 +1,2 @@
 export { DaffCartPaymentMethodIdMap } from './payment-method-map';
 export { DAFF_CART_ERROR_MATCHER } from './error-matcher.token';
-export { DaffCartItemStateDebounceTime } from './cart-item-state-debounce-time';
